feat: add AuthProvider
