+++ src/components/molecules/index.ts
@@ -12,3 +12,4 @@ export * from './CardsPreloader'
 export * from './ListsSkeleton'
 export * from './ListsItem'
 export * from './PageInfoBlock'
+export * from './AuthProvider'
