+++ src/templates/LayoutTemplate/LayoutTemplate.tsx
@@ -1,13 +1,9 @@
-import { ReactNode, FC, useEffect, useCallback } from 'react'
+import { ReactNode, FC } from 'react'
 import { Footer, Header, MobileMenu, Spacer } from '@ui-kit'
 import { Flexbox, PaddingContainer } from '@components/atoms'
 import { theme } from '@constants'
-import { EAuthorizationStatus, EPagePaths, THeaderFooter } from '@types'
-import { onAuthStateChanged } from 'firebase/auth'
-import { useAppDispatch } from '@hooks'
-import { setAuthStatus, setUser } from '@reducers'
-import { auth } from '@api'
-import { useRouter } from 'next/router'
+import { THeaderFooter } from '@types'
+import { AuthProvider } from '@components/molecules'
 
 type TLayoutTemplateProps = {
   children: ReactNode
@@ -15,42 +11,19 @@ type TLayoutTemplateProps = {
 }
 
 export const LayoutTemplate: FC<TLayoutTemplateProps> = ({ children, headerFooterData }) => {
-  const router = useRouter()
-  const { pathname, push } = router
   const { header, footer } = headerFooterData || {}
-  const dispatch = useAppDispatch()
-
-  const checkAuth = useCallback(async () => {
-    await onAuthStateChanged(auth, user => {
-      if (user) {
-        dispatch(setUser(user))
-        dispatch(setAuthStatus(EAuthorizationStatus.AUTH))
-
-        return
-      }
-
-      dispatch(setUser(null))
-      dispatch(setAuthStatus(EAuthorizationStatus.NO_AUTH))
-
-      if (pathname === EPagePaths.PROFILE || pathname === EPagePaths.MY_LISTS) {
-        push(EPagePaths.LOGIN)
-      }
-    })
-  }, [dispatch, pathname, push])
-
-  useEffect(() => {
-    checkAuth()
-  }, [checkAuth])
 
   return (
-    <Flexbox direction="column" height="100%">
-      <Header headerData={header} />
-      <PaddingContainer padding={theme.space.sm} mobOnly withMaxHeight>
-        {children}
-      </PaddingContainer>
-      <Spacer sizeMob={theme.space.xl4} size={0} />
-      <MobileMenu />
-      <Footer footerData={footer} />
-    </Flexbox>
+    <AuthProvider>
+      <Flexbox direction="column" height="100%">
+        <Header headerData={header} />
+        <PaddingContainer padding={theme.space.sm} mobOnly withMaxHeight>
+          {children}
+        </PaddingContainer>
+        <Spacer sizeMob={theme.space.xl4} size={0} />
+        <MobileMenu />
+        <Footer footerData={footer} />
+      </Flexbox>
+    </AuthProvider>
   )
 }
